Finish implementation of transformation from/to legacy JSON format
We can now create JSON from graphs and get the exact same result as what
what was passed as JSON to the GraphFactory. This can now be used for
fastapi to transform our model into a compatible JSON response and for
SQLALchemy.

Also fixed up a few funky compatibility issues:

* Outputs are non-empty, but connections are if no outputs are
  specified.
* `name` doesn't seem to exist for newly created workflows.

Also worked around a circular import.