Node __eq__ and __hash__ documentation fixes.
Removal of duplicated declaration of enable_multiple_edges_per_output in Module.

The doc of the check method in Node was rewritten and this method implementation and behaviour were changed
to check if all ports are in the valid range.

InconsistentEdgeError attributes names were changed to be in snake case.

Removal of PathWarnings.

Changed CyclicGraphError to store for each edge its ports and the algorithm for finding cycles was updated to initialize that extra data.

Node changed to store graph_id and the json to graph static factory was updated to initialize that value.

Fixed all subclasses of Node to be hashable and to use the hash and eql methods from Node.