Construct & deconstruct Trigger w/o raw_data
Everything is now derived from properties instead of a dict with
arbitrary values thrown into it.

Deduplicated a few of the properties by creating the intermediate class
WorkflowNode in the hierarchy.

Also gave up on the filters situation: some filters are actually the
EMPTY_FILTER, i.e.

    {
        "operator": "",
        "path": "",
        "selector": "",
        "value": "",
    }

however some modules expect

    [
       { text: selector, value: ''} {text: value, value: ''}
       {text: operator, value: ''} {text: path, value: ''}
    ]

No idea what's up with that. To my current knowledge, we don't even need
filters on this level and this is hence just the result of a suboptimal
abstraction.

Let's just ignore the case entirely, always fill in EMPTY_FILTER on
return and skip this when checking for symmetry.