Rework input filters & implement filter modules
First of all, there were a few things to be fixed in the input
filtering:

* `selector` is only there to modify a subsection of the workflow input.
  I.e. with a `selector` being `Event._AttributeFlattened.{n}`, this
  means that the list below this is supposed to be filtered, however the
  entire workflow data is supposed to be kept.

  The "backfilling" of filtered items is a little hairy:

  * An insert operation of X into `Foo.{n}.bar` inserts X into EACH list
    element below `Foo` 🤡

  * Hashmaps and Lists are represented by the same datastructure in PHP.
    You'll notice it because both things are treated equally. A bit of
    type juggling was necessary to simulate this.

  * We can't just go to the requested level and do `ptr = newvalue`:
    reassining in Python doesn't modify the pointer, but assigns the
    variable to another value.

    Instead, we usually check if we're about to reach the end (i.e. if
    only one element is left in the attribute path) and if that's the
    case we do `ptr[last_attribute] = newvalue`.

* The `in` / `not_in` actually need to be reversed. I.e. the operation
  is

    value in data[selector]

  This is why the filter module expects a single value only for
  in/not_in.

  Implemented `in_or` which accepts a list of possible values and checks
  if any of those is in the list below `data[selector]`.

Yes, the behavior is super-weird, but we want to accept arbitrary legacy
workflows and these may depend on this behavior implicitly.

Given the highly limited time, absolute lack of tests and observability in
the legacy system, I won't guarantee that this is a perfect replacement,
but a little closer at least.

Also implemented filter modules for adding / removing filters from
workflows. Got reminded that we want a dictionary for filters since the
filters are named. I.e., a filter called "A" can be added and _only
that_ filter can be dropped later on by a removal module.